requirements.txt
fastapi==0.95.2
numpy==1.24.3
Pillow==9.5.0
Pillow==9.5.0
Requests==2.31.0
streamlit==1.22.0
tensorflow==2.12.0
